refactor: 단일 이동, 멀티 이동 추상클래스 생성 및 적용
