+++ src/main/java/chess/domain/piece/PieceFactory.java
@@ -2,8 +2,13 @@ package chess.domain.piece;
 
 import chess.domain.Color;
 import chess.domain.Position;
+import chess.domain.piece.multiple.Bishop;
+import chess.domain.piece.multiple.Queen;
+import chess.domain.piece.multiple.Rook;
 import chess.domain.piece.pawn.BlackPawn;
 import chess.domain.piece.pawn.WhitePawn;
+import chess.domain.piece.single.King;
+import chess.domain.piece.single.Knight;
 import java.util.HashMap;
 import java.util.Map;
