+++ src/test/java/chess/domain/piece/multiple/QueenTest.java
@@ -1,13 +1,14 @@
-package chess.domain.piece;
+package chess.domain.piece.multiple;
 
 import static chess.domain.Color.BLACK;
 import static chess.domain.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import chess.domain.ChessBoard;
 import chess.domain.Color;
 import chess.domain.Position;
+import chess.domain.piece.Piece;
+import chess.domain.piece.multiple.Queen;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -16,41 +17,41 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.EnumSource;
 
-class RookTest {
+class QueenTest {
 
-    private Piece rook;
+    private Piece queen;
     private Position source;
 
     @BeforeEach
     void setUp() {
-        rook = new Rook(Color.WHITE);
+        queen = new Queen(Color.WHITE);
         source = new Position('d', '4');
     }
 
     @ParameterizedTest
     @EnumSource(Color.class)
-    @DisplayName("룩 기물 생성")
-    void createRook(Color color) {
-        assertThat(new Rook(color)).isInstanceOf(Rook.class);
+    @DisplayName("퀸 기물 생성")
+    void createQueen(Color color) {
+        assertThat(new Queen(color)).isInstanceOf(Queen.class);
     }
 
     @ParameterizedTest
-    @CsvSource(value = {"d,8,true", "d,7,true", "c,4,true", "c,5,false", "c,6,false"})
+    @CsvSource(value = {"a,7,true", "d,8,true", "h,8,true", "h,4,true", "e,6,false", "f,5,false", "a,8,false"})
     @DisplayName("목표 지점이 이동 가능 경로를 벗어났는지에 따른 이동 가능여부")
     void emptyDirection(char col, char row, boolean expected) {
         Position target = new Position(col, row);
-        ChessBoard chessBoard = new ChessBoard(Map.of(source, new Rook(WHITE)));
+        ChessBoard chessBoard = new ChessBoard(Map.of(source, new Queen(WHITE)));
 
-        assertThat(rook.isMovable(source, target, chessBoard)).isEqualTo(expected);
+        assertThat(queen.isMovable(source, target, chessBoard)).isEqualTo(expected);
     }
 
     @Test
     @DisplayName("가로막히지 않고 목표지점이 비어 있으면 이동 가능")
     void isMovableToClearEmptyPosition() {
         Position target = new Position('d', '8');
-        ChessBoard chessBoard = new ChessBoard(Map.of(source, new Rook(WHITE)));
+        ChessBoard chessBoard = new ChessBoard(Map.of(source, new Queen(WHITE)));
 
-        assertThat(rook.isMovable(source, target, chessBoard)).isTrue();
+        assertThat(queen.isMovable(source, target, chessBoard)).isTrue();
     }
 
     @ParameterizedTest
@@ -59,10 +60,10 @@ class RookTest {
     void isMovableToClearPiecePosition(Color color, boolean expected) {
         Position target = new Position('d', '8');
         ChessBoard chessBoard = new ChessBoard(Map.of(
-                source, new Rook(WHITE),
-                target, new Rook(color)));
+                source, new Queen(WHITE),
+                target, new Queen(color)));
 
-        assertThat(rook.isMovable(source, target, chessBoard)).isEqualTo(expected);
+        assertThat(queen.isMovable(source, target, chessBoard)).isEqualTo(expected);
     }
 
     @Test
@@ -70,10 +71,10 @@ class RookTest {
     void isMovableToNotClearEmptyPosition() {
         Position target = new Position('d', '8');
         ChessBoard chessBoard = new ChessBoard(Map.of(
-                source, new Rook(WHITE),
-                new Position('d', '7'), new Rook(BLACK)));
+                source, new Queen(WHITE),
+                new Position('d', '7'), new Queen(BLACK)));
 
-        assertThat(rook.isMovable(source, target, chessBoard)).isFalse();
+        assertThat(queen.isMovable(source, target, chessBoard)).isFalse();
     }
 
     @ParameterizedTest
@@ -82,10 +83,10 @@ class RookTest {
     void isMovableToNotClearPiecePosition(Color color) {
         Position target = new Position('d', '8');
         ChessBoard chessBoard = new ChessBoard(Map.of(
-                source, new Rook(WHITE),
-                new Position('d', '7'), new Rook(BLACK),
-                target, new Rook(color)));
+                source, new Queen(WHITE),
+                new Position('d', '7'), new Queen(BLACK),
+                target, new Queen(color)));
 
-        assertThat(rook.isMovable(source, target, chessBoard)).isFalse();
+        assertThat(queen.isMovable(source, target, chessBoard)).isFalse();
     }
 }
