+++ src/test/java/chess/domain/piece/single/KnightTest.java
@@ -1,4 +1,4 @@
-package chess.domain.piece;
+package chess.domain.piece.single;
 
 import static chess.domain.Color.WHITE;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -6,6 +6,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import chess.domain.ChessBoard;
 import chess.domain.Color;
 import chess.domain.Position;
+import chess.domain.piece.Piece;
+import chess.domain.piece.single.Knight;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
